Refactoring vterrors by moving content around. Naming the files according to what they should be used for. And removing unused code.
